Refactor secondary CC files into CC package
